Features search API includes extent of results
The features search API includes the extent (outer boundaries) of the
search results in the form [y1, x1, y2, x2], representing the lower left
and upper right corners of the boundaries of the results.

We want to zoom to the extent of the search results, so we have added
this to allow that functionality.

Because there is no hook in the pagination classes to allow additional
data fields to the response, we added the extent directly to the
response data. We attempted an extension of the pagination class to
allow for the addition of the extent data directly, but the current
setup of the classes involved didn't allow for this extension.