Add confirmed matches to canonical facility list when matching
Prior to this commit, when a contributor manually confirmed a facility match,
then uploaded a replacement list, the same potential match would be presented
for repeat confirmation because there was no mechanism to feed the confirmation
back into the matching process.

Our first attempt to resolve this was by augmenting the training data loading
from a static JSON file with previously confirmed matches, but this did not
boost the match scores enough.

This commit resolves the problem by augmenting the canonical list of facilities
with items pulled from confirmed matches. We generate temporary, "extended" IDs
for these items in the format `US2020052GKF19F_MATCH-23` and normalized them
back to real Facility IDs when we create model objects from the output of the
Dedupe matching process.