Query string created for polygonal search
Redux functionality has been added to support the polygonal search
feature. Given a list of points in the form of `[[lat, long]]` in the
redux state, a query string is created when searching for facilities
which filters the facilities search to only return facilities within
the polygon described by the given points.

In addition, the points listed in the URL query string are able to be
parsed in order to rehydrate the redux state on load.