Polygonal Search
Users are now able to draw a polygon on the map. The facility search
results are filtered to only include the facilities which are contained
within the boundaries of the polygon.

The polygon is rendered based on the boundaries filter, so it renders
correctly on page refreshes based on the query string in the URL. The
polygon is also removed when the user resets the filters, thereby
removing the boundaries.